backend/requirements.txt
Flask
flask-cors
numpy
pillow
tensorflow
pandas
